+++ assets/src/components/MethodCol.jsx
@@ -19,12 +19,12 @@ import { RecursionIcon } from "./RecursionIcon.jsx";
 import { methodHasDocumentation } from "../helpers/stateHelper.js";
 import { DocumentationIcon } from "./DocumentationIcon.jsx";
 import { Heading } from "./ui/heading.jsx";
+import { MacroIcon } from "./MacroIcon.jsx";
 
 function MethodType({ text, tooltip }) {
   const attributes = {
-    className: `method-type text-xs ${
-      text == "INT" ? "method-type-internal" : "method-ext"
-    }`,
+    className: `method-type text-xs ${text == "INT" ? "method-type-internal" : "method-ext"
+      }`,
   };
 
   return (
@@ -74,6 +74,9 @@ function MethodItem({ method, selectedMethod }) {
           </div>
         </GridListStart>
         <GridListSpacer />
+
+            {method.is_macro ? (<MacroIcon></MacroIcon>) : (<></>)}
+
         {hasDocumentation ? (
           <DocumentationIcon method={method}></DocumentationIcon>
         ) : (